refactor: 路銀ページを切り出し
